gcp/requirements.txt
tensorflow==2.6.0
google-cloud-storage==1.42.1
Pillow==8.2.0
Flask==2.0.1
fastapi==0.65.2
